Rename end to end tests to integration tests
